Maquina Expendedora

+++ src/modelo/Producto.java
@@ -18,7 +18,7 @@ public class Producto {
         return nombre;
     }
 
-    public double getPrecio() {
+    public int getPrecio() {
         return precio;
     }
